diabetes_app/requirements.txt
pandas
numpy
scikit-learn
catboost
shap
imbalanced-learn
pyreadstat
streamlit
openai
scikit-fuzzy
joblib
seaborn
networkx
matplotlib
azure-ai-inference
